pyproject.toml
[project]
name = "fds-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.6",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "seaborn>=0.13.2",
    "torch",
    "torchvision>=0.24.1",
    "wandb>=0.23.0",
]
